refactor: move initSentry and more to separate file
In order to keep the index.ts file minimal and focusing
on the public API.
